requirements.txt
Pillow==9.0.0
opencv-python-headless==4.6.0.66
numpy
pytesseract
requests
django
djangorestframework
gunicorn